backend/requirements.txt
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
phidata
langchain-groq
langchain
langchain_core
langchain_community
openai
pypdf
sqlalchemy
google-generativeai
boto3
crawl4ai
pydantic
pydantic-ai
pyautogen
dask[dataframe]
flaml[automl]
tavily-python
flask[async]
flask
flask-cors
uvicorn
streamlit
numpy
pandas
langchain-google-genai
typing
pydantic
langgraph
langchain-experimental
crewai
composio-gemini
mcp-use
lightrag-hku
sentence-transformers
pydantic-ai-slim==0.0.55
pydantic-evals==0.0.55
pydantic-graph==0.0.55
pydantic-settings==2.8.1
pydantic_core==2.33.1
arxiv
wikipedia
# vertexai==1.65.0
# langchain-google-vertexai==1.0.6
mcp[cli]
google-adk
stackapi
mlflow
pygithub